Renames TableDetail into TableDetailPage
