+++ amundsen_application/static/js/pages/TableDetailPage/ColumnList/styles.scss
