+++ amundsen_application/static/js/pages/TableDetailPage/ColumnListItem/styles.scss
