+++ amundsen_application/static/js/pages/TableDetailPage/ColumnStats/styles.scss
