+++ amundsen_application/static/js/pages/TableDetailPage/DataPreviewButton/constants.ts
