+++ amundsen_application/static/js/pages/TableDetailPage/DataPreviewButton/styles.scss
