+++ amundsen_application/static/js/pages/TableDetailPage/FrequentUsers/index.tsx
