+++ amundsen_application/static/js/pages/TableDetailPage/FrequentUsers/styles.scss
