+++ amundsen_application/static/js/pages/TableDetailPage/LineageLink/index.tsx
