+++ amundsen_application/static/js/pages/TableDetailPage/ReportTableIssue/index.tsx
