+++ amundsen_application/static/js/pages/TableDetailPage/ReportTableIssue/styles.scss
