+++ amundsen_application/static/js/pages/TableDetailPage/RequestDescriptionText/constants.ts
