+++ amundsen_application/static/js/pages/TableDetailPage/RequestDescriptionText/styles.scss
