+++ amundsen_application/static/js/pages/TableDetailPage/TableHeaderBullets/constants.ts
