+++ amundsen_application/static/js/pages/TableDetailPage/TableIssues/constants.ts
