+++ amundsen_application/static/js/pages/TableDetailPage/TableIssues/index.tsx
