+++ amundsen_application/static/js/pages/TableDetailPage/TableIssues/styles.scss
