+++ amundsen_application/static/js/pages/TableDetailPage/TableOwnerEditor/index.spec.tsx
