+++ amundsen_application/static/js/pages/TableDetailPage/WatermarkLabel/styles.scss
