+++ amundsen_application/static/js/pages/TableDetailPage/constants.ts
