+++ amundsen_application/static/js/pages/TableDetailPage/index.spec.tsx
