+++ amundsen_application/static/js/pages/TableDetailPage/index.tsx
