+++ amundsen_application/static/js/pages/TableDetailPage/styles.scss
